requirements.txt
google-genai==1.56.0
pyaudio
